Feat: 자식 컴포넌트 관리 추상화, beforeUnmount 라이프사이클 활성화
1. 컴포넌트가 key와 component-name태그를 가진 node에 대해 컴포넌트를 생성하도록 추상화.

2. 위의 동작을 위해 자식의 추가 제거를 추상화한 adjustChildComponents 함수 작성. 이 함수가 작동하면서 제거된 컴포넌트의 beforeUnmount를 호출하여 라이프사이클을 구현하도록 함.

4. 이제 template가 반환하는 HTML에서 컴포넌트의 target이 되는 node는 data-comonent-name과 data-key 어트리뷰트를 가지고있어야하며, data-key는 중복되면 안된다.

5. 이제 자식컴포넌트를 생성할 때 afterMount가 아닌 generateChildComponent를 작성해야하며, generateChildComponent는 name을 받아 해당하는 컴포넌트를 생성해 반환하도록 함.

+++ Vanilla JS로 웹 컴포넌트 만들기/src/App.mjs
@@ -1,7 +1,7 @@
 import Component from "./core/Component.mjs";
 import ItemAppender from "./components/ItemAppender.mjs";
-import Items from './components/Items.mjs';
-import ItemFilter from './components/ItemFilter.mjs';
+import Items from "./components/Items.mjs";
+import ItemFilter from "./components/ItemFilter.mjs";
 
 export default class App extends Component {
   setup() {
@@ -19,31 +19,41 @@ export default class App extends Component {
 
   template() {
     return `
-    <div class="itemAppender" component></div>
-    <div class="items" component></div>
-    <div class="itemFilter" component></div>
+    <div class="itemAppender" data-component-name="ItemAppender" data-key="1"></div>
+    <div class="items" data-component-name="Items" data-key="2"></div>
+    <div class="itemFilter" data-component-name="ItemFilter" data-key="3"></div>
     `;
   }
 
-  afterMount() {
-    this.childComponents = {
-      itemAppender: new ItemAppender(this.target.querySelector(".itemAppender"), () => { return {
-        addItem: this.addItem.bind(this),
-      }}),
-      items: new Items(this.target.querySelector(".items"), () => { return {
-        items: this.getFilteredItems(),
-        deleteItem: this.deleteItem.bind(this),
-        toggleItemFilter: this.toggleItemFilter.bind(this),
-      }}),
-      itemFilter: new ItemFilter(this.target.querySelector(".itemFilter"), () => { return {
-        filterItems: this.filterItems.bind(this),
-      }}),
+  generateChildComponent(name) {
+    if (name === "ItemAppender") {
+      return new ItemAppender(this.target.querySelector(".itemAppender"), () => {
+        return {
+          addItem: this.addItem.bind(this),
+        };
+      });
+    }
+    if (name === "Items") {
+      return new Items(this.target.querySelector(".items"), () => {
+        return {
+          items: this.getFilteredItems(),
+          deleteItem: this.deleteItem.bind(this),
+          toggleItemFilter: this.toggleItemFilter.bind(this),
+        };
+      });
+    }
+    if (name === "ItemFilter") {
+      return new ItemFilter(this.target.querySelector(".itemFilter"), () => {
+        return {
+          filterItems: this.filterItems.bind(this),
+        };
+      });
     }
   }
 
   getFilteredItems() {
     const { filter, items } = this.state;
-    if (filter) return items.filter(item => item.isFiltered === false);
+    if (filter) return items.filter((item) => item.isFiltered === false);
     return items;
   }
 
@@ -57,7 +67,10 @@ export default class App extends Component {
 
   deleteItem(index) {
     const items = [...this.state.items];
-    items.splice(items.findIndex(item => item.index === index), 1);
+    items.splice(
+      items.findIndex((item) => item.index === index),
+      1
+    );
     this.setState({ items });
   }
 
@@ -67,7 +80,7 @@ export default class App extends Component {
 
   toggleItemFilter(index) {
     const items = [...this.state.items];
-    const item = items.find(item => item.index === index);
+    const item = items.find((item) => item.index === index);
     item.isFiltered = !item.isFiltered;
     this.setState({ items });
   }
